+++ src/IntelliTect.Coalesce/Api/Controllers/ControllerHelpers.cs
@@ -1,6 +1,8 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Http;
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -22,6 +24,25 @@ namespace IntelliTect.Coalesce.Api.Controllers
                 if (closeUnderlying) stream.Dispose();
                 return memoryStream.ToArray();
             }
+        }
+
+        /// <summary>
+        /// Returns true if there was any attempt to pass in any value 
+        /// for the properties on the object named by <paramref name="prefix"/>. 
+        /// Returns false if the only key is <paramref name="prefix"/> itself with an empty value,
+        /// which is how null objects are serialized to formdata by coalesce-vue.
+        /// </summary>
+        public static bool HasAnyValue(this IFormCollection? form, string prefix)
+        {
+            if (form is null) return false;
+
+            var hasValue = form.Keys.Any(k => k.StartsWith(prefix + "[") || k.StartsWith(prefix + "."));
+
+            if (form[prefix] is [""] && !hasValue)
+            {
+                return false;
+            }
+            return true;
         }
     }
 }